refactor: 코드리뷰 코멘트 반영
