+++ src/app/(routes)/(tabs)/profile/today-quiz-setting/page.tsx
@@ -1,9 +1,10 @@
-import GoBackButton from '@/shared/components/go-back-button'
+import SetNote from '@/features/quiz/today-quiz-setting/components/set-note'
+import SetQuizCount from '@/features/quiz/today-quiz-setting/components/set-quiz-count'
+import { TodayQuizSettingProvider } from '@/features/quiz/today-quiz-setting/context/today-quiz-setting-context'
+import FixedBottom from '@/shared/components/custom/fixed-bottom'
+import GoBackButton from '@/shared/components/custom/go-back-button'
+import { Button } from '@/shared/components/ui/button'
 import Text from '@/shared/components/ui/text'
-import FixedBottom from '@/views/profile/today-quiz-setting/components/fixed-bottom'
-import SetNote from '@/views/profile/today-quiz-setting/components/set-note'
-import SetQuizCount from '@/views/profile/today-quiz-setting/components/set-quiz-count'
-import { TodayQuizSettingProvider } from '@/views/profile/today-quiz-setting/context/today-quiz-setting-context'
 
 const TodayQuizSettingPage = () => {
   return (
@@ -20,7 +21,14 @@ const TodayQuizSettingPage = () => {
 
         <SetNote />
 
-        <FixedBottom />
+        <FixedBottom className="bottom-[88px] right-1/2 flex translate-x-1/2 gap-[6px] pb-[36px]">
+          <Button variant={'largeRound'} colors={'tertiary'} className="w-[35%]">
+            초기화
+          </Button>
+          <Button variant={'largeRound'} colors={'primary'} className="w-[65%]">
+            저장하기
+          </Button>
+        </FixedBottom>
       </main>
     </TodayQuizSettingProvider>
   )
